+++ pmd-xml/src/test/java/net/sourceforge/pmd/lang/xml/LanguageVersionTest.java
@@ -8,9 +8,9 @@ import java.util.Arrays;
 import java.util.Collection;
 
 import net.sourceforge.pmd.AbstractLanguageVersionTest;
-import net.sourceforge.pmd.lang.pom.PomLanguageModule;
-import net.sourceforge.pmd.lang.wsdl.WsdlLanguageModule;
-import net.sourceforge.pmd.lang.xsl.XslLanguageModule;
+import net.sourceforge.pmd.lang.xml.pom.PomLanguageModule;
+import net.sourceforge.pmd.lang.xml.wsdl.WsdlLanguageModule;
+import net.sourceforge.pmd.lang.xml.xsl.XslLanguageModule;
 
 class LanguageVersionTest extends AbstractLanguageVersionTest {