+++ pmd-xml/src/test/java/net/sourceforge/pmd/lang/xml/PMDTaskTest.java
@@ -2,13 +2,16 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.ant;
+package net.sourceforge.pmd.lang.xml;
 
 import org.junit.jupiter.api.Test;
+
+import net.sourceforge.pmd.ant.AbstractAntTestHelper;
 
 class PMDTaskTest extends AbstractAntTestHelper {
 
     PMDTaskTest() {
+        super.pathToTestScript = "target/test-classes/net/sourceforge/pmd/lang/xml";
         super.antTestScriptFilename = "pmdtasktest.xml";
     }