+++ pmd-xml/src/test/java/net/sourceforge/pmd/lang/xml/xsl/rule/codestyle/UseConcatOnceTest.java
@@ -2,10 +2,10 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.xsl.rule.performance;
+package net.sourceforge.pmd.lang.xml.xsl.rule.codestyle;
 
 import net.sourceforge.pmd.testframework.PmdRuleTst;
 
-class AvoidAxisNavigationTest extends PmdRuleTst {
+class UseConcatOnceTest extends PmdRuleTst {
     // no additional unit tests
 }